Extract level attempt so it can be re-used in other tests
